+++ CHANGELOG.md
@@ -1,10 +1,29 @@
 # Changelog
 
+## 0.4.1
+- Added screenshot methods #14
+  ```dart
+  /// Takes a screenshot of the entire window
+  await takeScreenshot();
+  
+  /// Takes a screenshot of a single Screen/Widget
+  final homePage = spotSingle<HomePage>();
+  await takeScreenshot(selector: homePage);
+  
+  /// Use it as extension
+  await spotSingle<HomePage>().takeScreenshot();
+  ```
+- Export all types from `checks.dart` which are required to use `hasProp`
+- Update for Flutter 3.13
+
 ## 0.4.0
 - Added `act.tap(button)` to tap widgets #9
 - Raise min Flutter version to 3.10.0
 - Switch to the official [`checks`](https://pub.dev/packages/checks) package #12
 - Rename `SingleWidgetSnapshot.discoveredElements` -> `SingleWidgetSnapshot.discoveredElement` #11
+
+## 0.3.2
+- Export all types from `checks.dart` which are required to use `hasProp` 
 
 ## 0.3.1
 - Fix compilation error with Flutter 3.0.0